Fix issue uploading Image if none exists initially

+++ main_app/views.py
@@ -144,7 +144,8 @@ def add_photo(request, whey_id):
     photo_file = request.FILES.get('photo-file', None)
     if photo_file:
         # First delete previous photo (Only one photo is allowed)
-        Photo.objects.get(whey_id=whey_id).delete()
+        if (Photo.objects.filter(whey_id=whey_id)==True):
+            Photo.objects.get(whey_id=whey_id).delete()
         s3 = boto3.client('s3')
         # need a unique "key" for S3 / needs image file extension too
         key = uuid.uuid4().hex[:6] + photo_file.name[photo_file.name.rfind('.'):]
